+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - fixed `/tp` console command reporting teleport fails as success (#1484, regression from 4.1)
 - fixed console commands causing improper ring shutdown with selected inventory item (#1460, regression from 3.0)
 - fixed console input immediately ending demo (#1480, regression from 4.1)
+- changed `/tp` console command to look for the closest place to teleport to when targeting items (#1484)
 - improved level load times
 - improved logs module names readability
 - improved crash debug information on Windows